requirements.txt
Pillow==9.4.0
customtkinter==5.2.0